fix(codegen): 9 legacy-engine test failures were test-authoring bugs
Confirmed pre-existing at the wave base (3f43c7d4f) and unrelated to any
prepared-engine change this session made -- all 9 fixes are to the test
files only; emit/case.rs, emit/primop.rs and host_fns/primops.rs are
untouched, confirming the byte-array/float primop implementations were
already correct.

test_case_lit_float: built its LitFloat scrutinee/alt bit patterns via
f64::to_bits() instead of f32::to_bits() as u64, the engine's actual
documented Float# convention (emit/primop.rs's own unbox_float,
f32::from_bits(*bits as u32)). 1.0f64 and 2.5f64 truncate to the same
low-32-bits (0x0), so the case spuriously matched the first alt instead
of exercising the intended second one.

The other 8 (test_emit_primop_bytearray_{read_write, word_read_write,
set, shrink, index_word8, copy_addr, compare_unequal, copy_mutable}):
all sequenced a mutating PrimOp before an observing one via
`let dummy = <mutation> in <observe> + dummy`, relying on `dummy` being
forced before `<observe>` runs. GHC Core `let` is non-strict
(emit/expr.rs's own doc: "strictness is expressed via `case`, never
`let`"), and a PrimOp RHS is never trivial (tidepool_repr::trivial_field)
so it is always thunked -- two independently-thunked arguments to a
strict primop like IntAdd have no specified relative forcing order.
Confirmed empirically this observably forces the non-Var argument
first, so the read ran before the write's thunk was ever forced,
observing an unwritten array. Fixed by sequencing with `Case` instead,
whose scrutinee `emit_case::emit_case` unconditionally forces to WHNF as
its first step -- the actual GHC Core strictness idiom.

cargo nextest run -p tidepool-codegen --test codegen --no-fail-fast:
222/222 passed (2 slow), 2 skipped, 0 failed -- includes
apply_gc_during_application_relocates_forced_callee (fixed earlier this
session). cargo check -p tidepool-codegen --tests: clean.

+++ tidepool-codegen/tests/emit_case.rs
@@ -312,9 +312,18 @@ fn test_case_lit_double() {
 fn test_case_lit_float() {
     // case Lit(2.5f) of { LitAlt(1.0f) -> 10; LitAlt(2.5f) -> 77; Default -> 0 }
     let binder = VarId(99);
-    // LitFloat stores f64 bits (GHC represents Float# as Double internally)
-    let target_bits = 2.5f64.to_bits();
-    let one_bits = 1.0f64.to_bits();
+    // FINDING (pre-existing, stale expectation): `LitFloat` stores raw f32
+    // bits widened into a u64 -- see the engine's own convention at
+    // `Literal::LitFloat(1.5f32.to_bits() as u64)` a few tests below in this
+    // same file, `unbox_float` (tidepool-codegen/src/emit/primop.rs) which
+    // reinterprets the value slot as F32, and case.rs's own lit dispatch
+    // (`f32::from_bits(*bits as u32)`). This test previously stored f64 bits
+    // instead; truncating a f64 bit pattern to its low 32 bits collapses both
+    // 1.0f64 and 2.5f64 to 0x00000000 (both have a zero low word), so the
+    // scrutinee spuriously matched the *first* alt. Use f32 bits, matching
+    // the engine's actual Float# representation.
+    let target_bits = 2.5f32.to_bits() as u64;
+    let one_bits = 1.0f32.to_bits() as u64;
     let tree = RecursiveTree {
         nodes: vec![
             CoreFrame::Lit(Literal::LitFloat(target_bits)), // 0: scrutinee
